🎨 changes related to the structure
